refac: move tangent_lines into piecewise.py, remove linearization.py
Single function doesn't justify a separate module. tangent_lines
lives next to breakpoints() and segments() — all stateless helpers
for the piecewise workflow.
