+++ linopy/piecewise.py
@@ -359,6 +359,88 @@ def segments(
     return _coerce_segments(values, dim)
 
 
+def tangent_lines(
+    x: LinExprLike,
+    x_points: BreaksLike,
+    y_points: BreaksLike,
+) -> LinearExpression:
+    r"""
+    Compute tangent-line expressions for a piecewise linear function.
+
+    Returns a :class:`~linopy.expressions.LinearExpression` with an extra
+    segment dimension.  Each element along the segment dimension is the
+    tangent line of one segment: :math:`m_k \cdot x + c_k`.
+
+    Use the result in a regular constraint to create an upper or lower
+    bound:
+
+    .. code-block:: python
+
+        t = tangent_lines(power, x_pts, y_pts)
+        m.add_constraints(fuel <= t)  # upper bound (concave f)
+        m.add_constraints(fuel >= t)  # lower bound (convex f)
+
+    No auxiliary variables are created — the result is purely linear.
+
+    Parameters
+    ----------
+    x : Variable or LinearExpression
+        The input expression.
+    x_points : BreaksLike
+        Breakpoint x-coordinates (must be strictly increasing).
+    y_points : BreaksLike
+        Breakpoint y-coordinates.
+
+    Returns
+    -------
+    LinearExpression
+        Expression with an additional ``_breakpoint_seg`` dimension
+        (one entry per segment).
+    """
+    from linopy.expressions import LinearExpression as LinExpr
+    from linopy.variables import Variable
+
+    if not isinstance(x_points, DataArray):
+        x_points = _coerce_breaks(x_points)
+    if not isinstance(y_points, DataArray):
+        y_points = _coerce_breaks(y_points)
+
+    dx = x_points.diff(BREAKPOINT_DIM)
+    dy = y_points.diff(BREAKPOINT_DIM)
+    slopes = dy / dx
+
+    n_seg = slopes.sizes[BREAKPOINT_DIM]
+    seg_index = np.arange(n_seg)
+
+    slopes = slopes.rename({BREAKPOINT_DIM: LP_SEG_DIM})
+    slopes[LP_SEG_DIM] = seg_index
+
+    x_base = x_points.isel({BREAKPOINT_DIM: slice(None, -1)}).rename(
+        {BREAKPOINT_DIM: LP_SEG_DIM}
+    )
+    y_base = y_points.isel({BREAKPOINT_DIM: slice(None, -1)}).rename(
+        {BREAKPOINT_DIM: LP_SEG_DIM}
+    )
+    x_base[LP_SEG_DIM] = seg_index
+    y_base[LP_SEG_DIM] = seg_index
+
+    intercepts = y_base - slopes * x_base
+
+    if isinstance(x, Variable):
+        x_expr = x.to_linexpr()
+    elif isinstance(x, LinExpr):
+        x_expr = x
+    else:
+        raise TypeError(f"x must be a Variable or LinearExpression, got {type(x)}")
+
+    return slopes * x_expr + intercepts
+
+
+# ---------------------------------------------------------------------------
+# Internal validation
+# ---------------------------------------------------------------------------
+
+
 def _validate_xy_points(x_points: DataArray, y_points: DataArray) -> bool:
     """Validate x/y breakpoint arrays and return whether formulation is disjunctive."""
     if BREAKPOINT_DIM not in x_points.dims:
@@ -763,7 +845,7 @@ def add_piecewise_constraints(
     expressions.
 
     For inequality constraints (y <= f(x) or y >= f(x)), use
-    :func:`~linopy.linearization.tangent_lines` with regular
+    :func:`~linopy.piecewise.tangent_lines` with regular
     ``add_constraints`` instead.
 
     Example::
